tracks() now is expected to return a list of named tracks that whole data is stored in one chunk in .tracks.json using jsonpickle a track can now have a description that is displayed in the courses page course-show-tracks displays detailed tracks

+++ nbhosting/courses/management/commands/course_show_tracks.py
@@ -6,6 +6,8 @@ from django.core.management.base import BaseCommand
 
 from nbhosting.courses.model_course import CourseDir
 
+from nbhosting.main.settings import logger
+
 class Command(BaseCommand):
     help = 'Display static mappings for a course'
 
@@ -20,11 +22,16 @@ class Command(BaseCommand):
         preserve = kwargs['preserve']
         coursename = kwargs['coursename']
         coursedir = CourseDir(coursename)
+        logger.info(f"preserve mode = {preserve}")
         if not preserve:
-            caches = (coursedir.notebooks_dir / ".tracks").glob("*.json")
-            for cache in caches:
-                print(f"clearing cache {cache}")
+            cache = (coursedir.notebooks_dir / ".tracks.json")
+            if cache.exists():
+                logger.info(f"clearing cache {cache}")
                 cache.unlink()
         tracks = coursedir.tracks()
-        for trackname in tracks:
-            print(f"{trackname}: {coursedir.track(trackname)}")
+        for track in tracks:
+            logger.info(f"{track.name}: {track}")
+            for section in track.sections:
+                logger.info(f"{4*' '}{section.name}")
+                for notebook in section.notebooks:
+                    logger.info(f"{8*' '}{notebook.notebookname}")
